fix for unit-test
